+++ tests/qiskit_convert_test.py
@@ -1114,3 +1114,11 @@ def test_symbolic_param_conv() -> None:
             for i in range(len(qc_transpiled_again.parameters))
         }
     )
+
+
+# https://github.com/CQCL/pytket-qiskit/issues/337
+def test_nonregister_bits() -> None:
+    c = Circuit(1).X(0).measure_all()
+    c.rename_units({Bit(0): Bit(1)})
+    with pytest.raises(NotImplementedError):
+        tk_to_qiskit(c)