+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Unreleased
 * User can pass an `Options` instance (from `qiskit-ibm-runtime`)
   via a keyword argument to both an `IBMQBackend` constructor and 
   an instance method `IBMQBackend.process_circuits`.
+* Remove dependency on deprecated qiskit-ibm-provider.
 
 0.53.0 (April 2024)
 -------------------