+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Unreleased
   an instance method `IBMQBackend.process_circuits`.
 * Remove dependency on deprecated qiskit-ibm-provider.
 * Remove support for deprecated "ibmq_qasm_simulator" backend.
+* Forbid circuits with incomplete classical registers in ``tk_to_qiskit()``.
 
 0.53.0 (April 2024)
 -------------------