+++ docs/changelog.rst
@@ -10,6 +10,9 @@ Unreleased
 * Remove dependency on deprecated qiskit-ibm-provider.
 * Remove support for deprecated "ibmq_qasm_simulator" backend.
 * Forbid circuits with incomplete classical registers in ``tk_to_qiskit()``.
+* Updated pytket version requirement to 1.28.
+* Update qiskit-ibm-runtime version requirement to 0.24.
+* Update qiskit version requirement to 1.1.
 
 0.53.0 (April 2024)
 -------------------